Dumb warnings

+++ src/Basescape/SellState.cpp
@@ -267,7 +267,7 @@ void SellState::think()
  * @param selected Selected craft.
  * @return Index of the selected craft.
  */
-int SellState::getCraftIndex(size_t selected) const
+size_t SellState::getCraftIndex(size_t selected) const
 {
 	return selected - _soldiers.size();
 }
@@ -679,7 +679,7 @@ enum SellType SellState::getType(size_t selected) const
  * @param selected Currently selected item.
  * @return Index of the selected item.
  */
-int SellState::getItemIndex(size_t selected) const
+size_t SellState::getItemIndex(size_t selected) const
 {
 	return selected - _soldiers.size() - _crafts.size() - _hasSci - _hasEng;
 }
