+++ src/Ruleset/SoldierNamePool.cpp
@@ -124,7 +124,7 @@ std::wstring SoldierNamePool::genName(SoldierGender *gender) const
  * @param numLooks The maximum index.
  * @return The index of the soldier's look.
  */
-int SoldierNamePool::genLook(size_t numLooks)
+size_t SoldierNamePool::genLook(size_t numLooks)
 {
 	int look = 0;
 	const int minimumChance = 2;	// minimum chance of a look being selected if it isn't enumerated. This ensures that looks MUST be zeroed to not appear.
